[C8] Split tessellate_trim.go by mesher (#2224)
tessellate_trim.go was 860 LOC, 1.7x the 500-line ceiling. Split the
trimmed-surface mesher along its responsibilities, leaving the router in
place; each file is now under the ceiling:

- tessellate_trim.go (124): the router (tessellateCurvedFace,
  splineFaceMesh, meshSeamCrossingFace).
- tessellate_trim_grid.go (259): the structured (u,v) grid construction —
  iso rectangle, single- and doubly-periodic band grids, cell emission, and
  the grid-line tolerances/helpers.
- tessellate_trim_boundary.go (263): the trim-boundary -> (u,v) imprint
  (toUVLoops/toUVLoop/unwrap/seamWindingLeap/isPeriodic) and the boundary
  patch mesher (nonRectangularMesh/boundaryPatchMesh/projection).
- tessellate_trim_special.go (206): the special-case curved-mesher dispatch
  (specialCurvedMesh/specialCurvedMeshers) and the cone-apex fans.
- tessellate_trim_policy.go (75): the full-domain quality/deflection
  fallback (fullDomainGridMesh/unionIsoparmParams/mergeSortedParams).

Behaviour-preserving: every function body is byte-identical, only moved;
imports narrowed per file; the two grid tolerances kept their tol: notes.
The tessellation corpus, occtparity goldens (byte-identical, 307s), and
archguard pass before and after with no golden updates. Net delta of
recognizers / tolerance constants / fallback sites / type assertions: 0.

Note: the ordered dispatch table specialCurvedMeshers is moved verbatim
here; folding it into a classification is the separate phase-2 C1 task.

Part of #2183 (C8: structure/size limits). Closes #2224.

(cherry picked from commit f23fa347caced9f183e66b95bc74ac89682f5c10)
